Squashed commit of the following:
commit 89cf2441d50a56a912133cfca9ab12694c8e0857
Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
Date:   Fri Nov 22 22:04:58 2024 +0800

    brand deletion

commit 6ebf6143cae3ebeab8343eb2bc159307b53db024
Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
Date:   Fri Nov 22 21:31:19 2024 +0800

    Recent Changes

commit 85f84150934b937323127e18d63ea8345ae5d3da
Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
Date:   Fri Nov 22 20:40:25 2024 +0800

    revised changes

commit 5d77c3a9ac21f6a499089931a6a2426516562719
Merge: 13d01c8 a708526
Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
Date:   Fri Nov 22 20:31:03 2024 +0800

    changes

commit 13d01c81d42dc90fd370f4da2e6f728afa131803
Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
Date:   Fri Nov 22 20:16:03 2024 +0800

    Squashed commit of the following:

    commit a708526de4e40ae20103ba83acc8c9d2771aa9af
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Fri Nov 22 18:36:26 2024 +0800

        cart

    commit 0b3fe46205fe33b9ff3a13053bc893cdb89dca47
    Merge: 5909c37 6fd8087
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Fri Nov 22 10:23:22 2024 +0800

        Merge branch 'main' of https://github.com/ChristInsanity/SupplyBridge

    commit 5909c37c7f15e0579b01ba69d389fc0570d33047
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Fri Nov 22 09:24:50 2024 +0800

        s

    commit 0f9121838d18ea1056260c45957b9315dcde9849
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Thu Nov 21 19:03:41 2024 +0800

        Shop2

    commit 5b6d04871b924f9d5717d1df986991d001238710
    Merge: fdc3782 f7803e4
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Thu Nov 21 00:14:46 2024 +0800

        Merge remote-tracking branch 'upstream/main'

    commit fdc37825925aef049d3de9f2111bff67189b146a
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Thu Nov 21 00:12:22 2024 +0800

        shop

    commit 3b5e2b40ed819c2cb7cdd87de1e657f1710cb960
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Wed Nov 20 22:15:55 2024 +0800

        finally

commit b1c114f0369ba5795ec7e4634c26dc0a72c310f8
Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
Date:   Fri Nov 22 19:59:22 2024 +0800

    troubleshooting for recent bug

commit 54a0d4811da1e855a0f5e26b9c1922569ed095db
Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
Date:   Fri Nov 22 15:57:50 2024 +0800

    Squashed commit of the following:

    commit dea3a839dd48a6b68db5c7ac4f2c174f791c645e
    Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
    Date:   Fri Nov 22 15:52:08 2024 +0800

        CategoryTroubleshooting

    commit 27a460d5cc14b230cb899a373b8a65ce41697deb
    Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
    Date:   Fri Nov 22 10:11:06 2024 +0800

        Squashed commit of the following:

        commit 58425f754e959bc88989526bb7b1460d39d6225d
        Author: aj-u3u <arjeanlagrosa2004@gmail.com>
        Date:   Fri Nov 22 08:50:16 2024 +0800

            updated dev dependencies

            dev dependencies version fixed

        commit 6e8d5adad364d0a2e9fa651aabc329226f98bae8
        Author: aj-u3u <arjeanlagrosa2004@gmail.com>
        Date:   Fri Nov 22 08:00:53 2024 +0800

            picture

            changes in picture in index

commit a8cde4409ff5b63ec441259e59f6bb350b58c4c9
Author: John Lyold Lozada <167066974+DeJure12705@users.noreply.github.com>
Date:   Fri Nov 22 10:20:44 2024 +0800

    Squashed commit of the following:

    commit 5909c37c7f15e0579b01ba69d389fc0570d33047
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Fri Nov 22 09:24:50 2024 +0800

        s

    commit 0f9121838d18ea1056260c45957b9315dcde9849
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Thu Nov 21 19:03:41 2024 +0800

        Shop2

    commit 5b6d04871b924f9d5717d1df986991d001238710
    Merge: fdc3782 f7803e4
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Thu Nov 21 00:14:46 2024 +0800

        Merge remote-tracking branch 'upstream/main'

    commit fdc37825925aef049d3de9f2111bff67189b146a
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Thu Nov 21 00:12:22 2024 +0800

        shop

    commit 3b5e2b40ed819c2cb7cdd87de1e657f1710cb960
    Author: ChristInsanity <xian23xian23xian@gmail.com>
    Date:   Wed Nov 20 22:15:55 2024 +0800

        finally

+++ app/Http/Controllers/AdminController.php
@@ -21,8 +21,8 @@ class AdminController extends Controller
 
     public function brands()
     {
-        $brands = Brand::orderBy('id', 'DESC')->paginate(10);
-        return view('admin.brands', compact('brands'));
+       $brands = Brand::orderBy('id', 'DESC')->paginate(10);
+        return view('admin.brands', compact( 'brands'));
     }
 
     public function add_Brand()
@@ -52,14 +52,16 @@ class AdminController extends Controller
     public function GenerateBrandThumbnailsImage($image, $imageName)
     {
         $destinationPath = public_path('uploads/brands');
+        
         if (!File::exists($destinationPath)) {
             File::makeDirectory($destinationPath, 0755, true);
         }
         $img = Image::read($image->path());
-        $img->cover(124,124,"top");
-        $img->resize(124,124, function($constraint){
+        $img->cover(124, 124, "top");
+        $img->resize(124, 124,function($constraint){
             $constraint->aspectRatio();
-        })->save($destinationPath.'/'.$imageName);
+        });
+        $img->save($destinationPath.'/'.$imageName);
     }
     public function brand_edit($id)
     {
@@ -84,7 +86,6 @@ class AdminController extends Controller
     $brand->slug = Str::slug($request->name);
 
     if ($request->hasFile('image')) {
-        // Delete old image if exists
         if (File::exists(public_path('uploads/brands').'/'.$brand->image)) {
             File::delete(public_path('uploads/brands').'/'.$brand->image);
         }
@@ -98,65 +99,65 @@ class AdminController extends Controller
     $brand->save();
     return redirect()->route('admin.brands')->with('success', 'Brand updated successfully');
 }
+public function delete_brand($id)
+    {
+        $brand = Brand::find($id);
+        if (!$brand) {
+            return redirect()->route('admin.brands')->with('error', 'Brand not found.');
+        }
+
+        if (File::exists(public_path('uploads/brands').'/'.$brand->image)) {
+            File::delete(public_path('uploads/brands').'/'.$brand->image);
+        }
+
+        $brand->delete();
+        return redirect()->route('admin.brands')->with('success', 'Brand deleted successfully.');
+    }
 
     
 
     public function categories(){
-        try {
-            $categories = Category::orderBy('id', 'DESC')->paginate(10);
-            return view('admin.categories', compact('categories'));
-        } catch (\Exception $e) {
-            Log::error('Error fetching categories: '.$e->getMessage());
-            return redirect()->back()->with('error', 'An error occurred while fetching the categories.');
-        }
+        
+      $categories = Category::orderBy('id', 'DESC')->paginate(10);
+      return view('admin.categories', compact('categories'));
+
     }
     public function category_add(){
         return view('admin.category-add');
     }
     public function category_store(Request $request)
     {
-        try {
+       
             $request->validate([
                 'name' => 'required',
                 'slug' => 'required|unique:categories,slug',
                 'image' => 'mimes:png,jpg,jpeg|max:2048',
             ]);
 
-            $category = new Category();
-            $category->name = $request->name;
-            $category->slug = Str::slug($request->name);
-
-            if ($request->hasFile('image')) {
+                $category = new Category();
+                $category->name = $request->name;
+                $category->slug = Str::slug($request->name);
                 $image = $request->file('image');
                 $file_extention = $image->extension();
                 $file_name = Carbon::now()->timestamp.'.'.$file_extention;
-                $this->GenerateCategoriesThumbnailsImage($image, $file_name);
+                $this->GenerateCategoryThumbnailsImage($image, $file_name);
                 $category->image = $file_name;
+                $category->save();
+                return redirect()->route('admin.categories')->with('success', 'Category added successfully');
+                
             }
 
-            $category->save();
-            return redirect()->route('admin.categories')->with('success', 'Category added successfully');
-        } catch (\Exception $e) {
-            Log::error('Error adding category: '.$e->getMessage());
-            return redirect()->back()->with('error', 'An error occurred while adding the category.');
-        }
-    }
+      
+    
 
-    public function GenerateCategoriesThumbnailsImage($image, $imageName)
+    public function GenerateCategoryThumbnailsImage($image, $imageName)
     {
-        try {
-            $destinationPath = public_path('uploads/categories');
-            if (!File::exists($destinationPath)) {
-                File::makeDirectory($destinationPath, 0755, true);
-            }
-            $img = Image::read($image->path()); // Ensure this line is correct
-            $img->resize(124, 124, function($constraint){
-                $constraint->aspectRatio();
-            })->save($destinationPath.'/'.$imageName);
-        } catch (\Exception $e) {
-            Log::error('Error generating category thumbnail: '.$e->getMessage());
-            throw $e;
-        }
+        $destinationPath = public_path('uploads/categories');
+        $img = Image::read($image->path());
+        $img->cover(124, 124, "top");
+        $img->resize(124, 124);
+        $img->save($destinationPath.'/'.$imageName);
+    
     }
     public function category_edit($id)
     {
@@ -166,29 +167,35 @@ class AdminController extends Controller
         
     }
     public function category_update(Request $request){
-        $request->validate([
-            'name' => 'required',
-            'slug' => 'required|unique:categories,slug'.$request->id,
-            'image' => 'mimes:png,jpg,jpeg|max:2048',
-        ]);
-
-        $category = Category::find($request->id);
-        $category->name = $request->name;
-        $category->slug = Str::slug($request->name);
-        if ($request->hasFile('image')) {
-            if(File::exists(public_path('uploads/categories').'/'.$category->image)){
-                File::delete(public_path('uploads/categories').'/'.$category->image);
+        
+            $request->validate([
+                'name' => 'required',
+                'slug' => 'required|unique:categories,slug,'.$request->id,
+                'image' => 'mimes:png,jpg,jpeg|max:2048',
+            ]);
+        
+            $category = Category::find($request->id);
+            if (!$category) {
+                return redirect()->back()->with('error', 'Category not found.');
             }
-            $image = $request->file('image');
-            $file_extention = $image->extension();
-            $file_name = Carbon::now()->timestamp.'.'.$file_extention;
-            $this->GenerateCategoriesThumbnailsImage($image, $file_name);
-            $category->image = $file_name;
-        }
-        $category->save();
-        return redirect()->route('admin.categories')->with('status', 'Category has been updated successfully'); 
-
-
+        
+            $category->name = $request->name;
+            $category->slug = Str::slug($request->name);
+        
+            if ($request->hasFile('image')) {
+                if (File::exists(public_path('uploads/categories').'/'.$category->image)) {
+                    File::delete(public_path('uploads/categories').'/'.$category->image);
+                }
+        
+                $image = $request->file('image');
+                $file_name = Carbon::now()->timestamp.'.'.$image->extension();
+                $this->GenerateCategoryThumbnailsImage($image, $file_name);
+                $category->image = $file_name;
+            }
+        
+            $category->save();
+            return redirect()->route('admin.categories')->with('success', 'Category updated successfully');
+        
     }
     public function category_delete($id){
         $category = Category::find($id);
@@ -273,20 +280,32 @@ class AdminController extends Controller
         return redirect()->route('admin.products')->with('success', 'Product has been added successfully');
     }
   
-  public function GenerateProductThumbnailImage($image, $imageName){
-    $destinationPathThumbnail = public_path('uploads/products/thumbnails');
-    $destinationPath = public_path('uploads/products');
-    $img = Image::read($image->path());
-
-    $img->cover(540, 689,'top');
-    $img->resize(540, 689, function ($constraint) {
-        $constraint->aspectRatio();
-    })->save($destinationPath.'/'.$imageName);
-
-    $img->resize(540, 689, function ($constraint) {
-        $constraint->aspectRatio();
-    })->save($destinationPathThumbnail.'/'.$imageName);
-  }
+    public function GenerateProductThumbnailImage($image, $imageName)
+    {
+        // Set paths
+        $destinationPathThumbnail = public_path('uploads/products/thumbnails');
+        $destinationPath = public_path('uploads/products');
+    
+        // Create directories if they don't exist
+        if (!File::exists($destinationPath)) {
+            File::makeDirectory($destinationPath, 0755, true);
+        }
+        if (!File::exists($destinationPathThumbnail)) {
+            File::makeDirectory($destinationPathThumbnail, 0755, true);
+        }
+    
+        // Process main image
+        $img = Image::read($image->path());
+        $img->cover(540, 689, 'top');
+        $img->save($destinationPath . '/' . $imageName);
+    
+        // Create thumbnail from original image
+        $thumbnail = Image::read($image->path());
+        $thumbnail->cover(540, 689, 'top');
+        $thumbnail->save($destinationPathThumbnail . '/' . $imageName);
+    
+        return true;
+    }
 
   public function product_edit($id)
   {
